Add files via upload
Here is my "final" product
I was under the impression that we could use as many of the files as we wanted to. 
I was working with my instructor to try to publish my dashboard to Tableau Public but I had trouble because of the size of my file.
My vizualizations are about 2020 data.